moved git formatter
