+++ packages/deprecation-crawler/src/lib/output-formatters/index.ts
@@ -1,4 +1,4 @@
-export {addCommentToRepository} from './add-comments-to-repository';
+export {addCommentToRepository} from './git/add-comments-to-repository';
 export {generateRawJson} from './json/raw.json.formatter';
 import {generateTagBasedFormat} from './markdown/tag-based.md-formatter';
 import {generateGroupBasedFormat} from './markdown/group-based.md.formatter';
